Revert "Merge pull request #1072 from layer2tech/feat/rejoin_autoswap_test"
This reverts commit 5b9122378354b469f85c468e5184ea43cae2fa76, reversing
changes made to 1f955c5a0354c655887e94dc12a312320ce774ec.
