+++ package.json
@@ -52,7 +52,6 @@
     "app-dev": "concurrently \"BROWSER=none yarn start\" \"wait-on http://localhost:3000 && electron .\""
   },
   "devDependencies": {
-    "@testing-library/react": "^12.1.2",
     "@types/uuid": "^8.3.0",
     "concurrently": "^5.3.0",
     "electron": "^10.1.4",
