+++ src/containers/CreateWalletInfo/CreateWalletInfo.js
@@ -7,13 +7,7 @@ import {Link, withRouter} from "react-router-dom";
 import './CreateWalletInfo.css'
 import { callGetArgsHasTestnet } from '../../features/WalletDataSlice';
 
-
-let Store
-try {
-  Store = window.require('electron-store');
-} catch (e) {
-  Store = require('electron-store');
-}
+let Store = window.require('electron-store');
 new Store();
 
 const CreateWalletInfoPage = () => {
